requirements.txt
blessings==1.7
bpython==0.17.1
certifi==2018.10.15
chardet==3.0.4
curtsies==0.3.0
Django==2.1.2
django-polymorphic==2.0.3
djangorestframework==3.9.0
greenlet==0.4.15
idna==2.7
Markdown==3.0.1
Pygments==2.2.0
pytz==2018.5
requests==2.20.0
six==1.11.0
typing==3.6.6
urllib3==1.24
wcwidth==0.1.7
